fixes for gcc 4.3
git-svn-id: http://svn.cern.ch/guest/lhcb/LHCb/trunk@60044 4525493e-7705-40b1-a816-d608a930855b

Former-commit-id: 236bd6c3a37fb45f6197e7c74d24ee7e5f6e1488

+++ LHCb/L0/ProcessorKernel/src/TileRegister.cpp
@@ -135,8 +135,8 @@ void L0Muon::TileRegister::print_bits(unsigned int event, FILE *file){
     if (val) {
       
       if (mid.isValid()){
-        fprintf(file,"%5u %3d ( %2d,%2d ) Q%1d M%1d R%1d %2d %2d  %d\n"
-                ,event ,i 
+        fprintf(file,"%5u %3u ( %2d,%2d ) Q%1d M%1d R%1d %2d %2d  %d\n"
+                ,event ,(unsigned long)i 
                 ,mid.layout().xGrid()
                 ,mid.layout().yGrid()
                 ,mid.quarter()+1 
@@ -147,7 +147,7 @@ void L0Muon::TileRegister::print_bits(unsigned int event, FILE *file){
                 ,val
                 ); 
       } else {
-        fprintf(file,"%5u %3d                           %d\n",event,i,val);
+        fprintf(file,"%5u %3u                           %d\n",event,i,val);
       }
     }
